requirements.txt
Django==1.9.7
djangorestframework==3.3.3
gevent==1.1.1
greenlet==0.4.10
grequests==0.3.0
jsonfield==1.0.3
numpy==1.11.1
pandas==0.18.1
plotly==1.12.4
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.10.0
six==1.10.0
wheel==0.24.0
